# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social Account Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface HealthStatus {
  status: string;
}

export type AccountPlatform = typeof AccountPlatform[keyof typeof AccountPlatform];


export const AccountPlatform = {
  facebook: 'facebook',
  instagram: 'instagram',
} as const;

export interface Account {
  id: number;
  platform: AccountPlatform;
  nama: string;
  /** @nullable */
  tgl_lahir?: string | null;
  /** @nullable */
  jenis_kelamin?: string | null;
  email: string;
  sandi: string;
  /** @nullable */
  sandi_fb?: string | null;
  /** @nullable */
  kode_2fa?: string | null;
  /** @nullable */
  uid?: string | null;
  /** @nullable */
  sandi_email?: string | null;
  /** @nullable */
  email_pemulihan?: string | null;
  status: string;
  /** @nullable */
  catatan?: string | null;
  createdAt: string;
}

export type AccountInputPlatform = typeof AccountInputPlatform[keyof typeof AccountInputPlatform];


export const AccountInputPlatform = {
  facebook: 'facebook',
  instagram: 'instagram',
} as const;

export interface AccountInput {
  platform: AccountInputPlatform;
  nama: string;
  tgl_lahir?: string;
  jenis_kelamin?: string;
  email: string;
  sandi: string;
  sandi_fb?: string;
  kode_2fa?: string;
  uid?: string;
  sandi_email?: string;
  email_pemulihan?: string;
  status?: string;
  catatan?: string;
}

export type AccountUpdatePlatform = typeof AccountUpdatePlatform[keyof typeof AccountUpdatePlatform];


export const AccountUpdatePlatform = {
  facebook: 'facebook',
  instagram: 'instagram',
} as const;

export interface AccountUpdate {
  platform?: AccountUpdatePlatform;
  nama?: string;
  tgl_lahir?: string;
  jenis_kelamin?: string;
  email?: string;
  sandi?: string;
  sandi_fb?: string;
  kode_2fa?: string;
  uid?: string;
  sandi_email?: string;
  email_pemulihan?: string;
  status?: string;
  catatan?: string;
}

export type AccountStatsByStatusItem = {
  status: string;
  count: number;
};

export interface AccountStats {
  total: number;
  facebook: number;
  instagram: number;
  byStatus: AccountStatsByStatusItem[];
}

export interface AccountImportRequest {
  accounts: AccountInput[];
}

export type AccountImportResultErrorsItem = {
  row: number;
  error: string;
};

export interface AccountImportResult {
  imported: number;
  failed: number;
  total: number;
  errors?: AccountImportResultErrorsItem[];
}

export interface ErrorResponse {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListAccountsParams = {
platform?: ListAccountsPlatform;
status?: string;
};

export type ListAccountsPlatform = typeof ListAccountsPlatform[keyof typeof ListAccountsPlatform];


export const ListAccountsPlatform = {
  facebook: 'facebook',
  instagram: 'instagram',
} as const;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
};