# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Social Account Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: accounts
    description: Social media account management
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [facebook, instagram]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /accounts/import:
    post:
      operationId: importAccounts
      tags: [accounts]
      summary: Bulk import accounts from CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountImportRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountImportResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /accounts/stats:
    get:
      operationId: getAccountStats
      tags: [accounts]
      summary: Get account statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStats"

  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a single account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Account:
      type: object
      required:
        - id
        - platform
        - nama
        - email
        - sandi
        - status
        - createdAt
      properties:
        id:
          type: integer
        platform:
          type: string
          enum: [facebook, instagram]
        nama:
          type: string
        tgl_lahir:
          type: ["string", "null"]
        jenis_kelamin:
          type: ["string", "null"]
        email:
          type: string
        sandi:
          type: string
        sandi_fb:
          type: ["string", "null"]
        kode_2fa:
          type: ["string", "null"]
        uid:
          type: ["string", "null"]
        sandi_email:
          type: ["string", "null"]
        email_pemulihan:
          type: ["string", "null"]
        status:
          type: string
        catatan:
          type: ["string", "null"]
        createdAt:
          type: string

    AccountInput:
      type: object
      required:
        - platform
        - nama
        - email
        - sandi
      properties:
        platform:
          type: string
          enum: [facebook, instagram]
        nama:
          type: string
        tgl_lahir:
          type: string
        jenis_kelamin:
          type: string
        email:
          type: string
        sandi:
          type: string
        sandi_fb:
          type: string
        kode_2fa:
          type: string
        uid:
          type: string
        sandi_email:
          type: string
        email_pemulihan:
          type: string
        status:
          type: string
        catatan:
          type: string

    AccountUpdate:
      type: object
      properties:
        platform:
          type: string
          enum: [facebook, instagram]
        nama:
          type: string
        tgl_lahir:
          type: string
        jenis_kelamin:
          type: string
        email:
          type: string
        sandi:
          type: string
        sandi_fb:
          type: string
        kode_2fa:
          type: string
        uid:
          type: string
        sandi_email:
          type: string
        email_pemulihan:
          type: string
        status:
          type: string
        catatan:
          type: string

    AccountStats:
      type: object
      required:
        - total
        - facebook
        - instagram
        - byStatus
      properties:
        total:
          type: integer
        facebook:
          type: integer
        instagram:
          type: integer
        byStatus:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                type: string
              count:
                type: integer

    AccountImportRequest:
      type: object
      required: [accounts]
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountInput"

    AccountImportResult:
      type: object
      required: [imported, failed, total]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        errors:
          type: array
          items:
            type: object
            required: [row, error]
            properties:
              row:
                type: integer
              error:
                type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string