# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social Account Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountPlatform } from './accountPlatform';

export interface Account {
  id: number;
  platform: AccountPlatform;
  nama: string;
  /** @nullable */
  tgl_lahir?: string | null;
  /** @nullable */
  jenis_kelamin?: string | null;
  email: string;
  sandi: string;
  /** @nullable */
  sandi_fb?: string | null;
  /** @nullable */
  kode_2fa?: string | null;
  /** @nullable */
  uid?: string | null;
  /** @nullable */
  sandi_email?: string | null;
  /** @nullable */
  email_pemulihan?: string | null;
  status: string;
  /** @nullable */
  catatan?: string | null;
  createdAt: string;
}