# lib/api-zod/src/generated/types/accountStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social Account Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountStatsByStatusItem } from './accountStatsByStatusItem';

export interface AccountStats {
  total: number;
  facebook: number;
  instagram: number;
  byStatus: AccountStatsByStatusItem[];
}