# lib/api-zod/src/generated/types/accountUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social Account Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountUpdatePlatform } from './accountUpdatePlatform';

export interface AccountUpdate {
  platform?: AccountUpdatePlatform;
  nama?: string;
  tgl_lahir?: string;
  jenis_kelamin?: string;
  email?: string;
  sandi?: string;
  sandi_fb?: string;
  kode_2fa?: string;
  uid?: string;
  sandi_email?: string;
  email_pemulihan?: string;
  status?: string;
  catatan?: string;
}